refactor : enum 객체(Message, GameStatus) 별도 분리
