+++ src/main/java/GameApplication.java
@@ -1,11 +1,12 @@
-import domain.Game;
+import domain.game.Game;
+import domain.game.GameStatus;
 
 public class GameApplication {
 
     public static void main(String[] args){
 
         Game game = new Game();
-        while(game.getStatus() == Game.Status.ONGOING){
+        while(game.getStatus() == GameStatus.ONGOING){
             game.play();
         }
     }
